+++ PlayerMicroservice/src/main/java/org/player/entity/Player.java
@@ -19,6 +19,6 @@ public class Player {
     private String playerPhone;
     private String playerPicture;
     private int playerTotalXP;
-    private int playerLevel;
+    private byte playerLevel;
     private boolean playerCanDoExam;
 }
